solutions/stage4/sources-ocr/requirements.txt
celery==5.2.3
Pillow==9.0.1